renderDocFileName in utils
