+++ apps/consultation-portal/utils/helpers/index.ts
@@ -29,3 +29,4 @@ export * from './createUUIDString'
 export * from './useIsMobile'
 export * from './useDebounce'
 export * from './debounce'
+export * from './renderDocFileName'
